[RHTAP-1011] Changed dir naming
