+++ pkg/utils/common/pod.go
@@ -3,8 +3,10 @@ package common
 import (
 	"context"
 	"fmt"
+	"os"
 	"time"
 
+	. "github.com/onsi/ginkgo/v2"
 	"github.com/redhat-appstudio/e2e-tests/pkg/utils"
 	"github.com/tektoncd/pipeline/pkg/apis/pipeline/v1beta1"
 	corev1 "k8s.io/api/core/v1"
@@ -86,6 +88,39 @@ func (s *SuiteController) WaitForPodSelector(
 	for i := range podList.Items {
 		if err := utils.WaitUntil(fn(podList.Items[i].Name, namespace), time.Duration(timeout)*time.Second); err != nil {
 			return err
+		}
+	}
+	return nil
+}
+
+// Stores logs of all pods of the job
+func (cs *SuiteController) StorePodLogs(testNamespace, jobName, testLogsDir string) error {
+	if err := os.MkdirAll(testLogsDir, os.ModePerm); err != nil {
+		return err
+	}
+
+	podList, err := cs.KubeInterface().CoreV1().Pods(jobName).List(context.TODO(), metav1.ListOptions{})
+	if err != nil {
+		return fmt.Errorf("error listing %s pods: %s\n", testNamespace, err.Error())
+	}
+
+	GinkgoWriter.Printf("found %d pods in namespace: %s\n", len(podList.Items), testNamespace)
+
+	for _, pod := range podList.Items {
+		var containers []corev1.Container
+		containers = append(containers, pod.Spec.InitContainers...)
+		containers = append(containers, pod.Spec.Containers...)
+		for _, c := range containers {
+			log, err := utils.GetContainerLogs(cs.KubeInterface(), pod.Name, c.Name, pod.Namespace)
+			if err != nil {
+				GinkgoWriter.Printf("error getting logs for pod/container %s/%s: %v\n", pod.Name, c.Name, err.Error())
+				continue
+			}
+
+			filepath := fmt.Sprintf("%s/%s-pod-%s-%s.log", testLogsDir, pod.Namespace, pod.Name, c.Name)
+			if err := os.WriteFile(filepath, []byte(log), 0644); err != nil {
+				return err
+			}
 		}
 	}
 	return nil
