+++ pkg/utils/tekton/pipelines.go
@@ -7,6 +7,7 @@ import (
 	"strconv"
 
 	. "github.com/onsi/ginkgo/v2"
+	g "github.com/onsi/ginkgo/v2"
 
 	app "github.com/redhat-appstudio/application-api/api/v1alpha1"
 	"github.com/redhat-appstudio/e2e-tests/pkg/utils"
@@ -228,6 +229,43 @@ func StorePipelineRun(pipelineRun *v1beta1.PipelineRun, c crclient.Client, ki ku
 		if err := os.WriteFile(filePath, pipelineRunYaml, 0644); err != nil {
 			GinkgoWriter.Printf("cannot write to %s: %+v\n", filename, err)
 			GinkgoWriter.Printf("\n%s\nFailed pipelineRunYaml:\n%s\n", filename, pipelineRunYaml)
+		}
+	}
+
+	return nil
+}
+
+func (s *SuiteController) StorePipelineRuns(componentPipelineRun *v1beta1.PipelineRun, testLogsDir, testNamespace string) error {
+	if err := os.MkdirAll(testLogsDir, os.ModePerm); err != nil {
+		return err
+	}
+
+	filepath := fmt.Sprintf("%s/%s-pr-%s.log", testLogsDir, testNamespace, componentPipelineRun.Name)
+	pipelineLogs, err := s.GetPipelineRunLogs(componentPipelineRun.Name, testNamespace)
+	if err != nil {
+		g.GinkgoWriter.Printf("got error fetching PR logs: %v\n", err.Error())
+	} else {
+		if err := os.WriteFile(filepath, []byte(pipelineLogs), 0644); err != nil {
+			g.GinkgoWriter.Printf("cannot write to %s: %+v\n", filepath, err)
+		}
+	}
+
+	pipelineRuns, err := s.ListAllPipelineRuns(testNamespace)
+
+	if err != nil {
+		return fmt.Errorf("got error fetching PR list: %v\n", err.Error())
+	}
+
+	for _, pipelineRun := range pipelineRuns.Items {
+		filepath := fmt.Sprintf("%s/%s-pr-%s.log", testLogsDir, testNamespace, pipelineRun.Name)
+		pipelineLogs, err := s.GetPipelineRunLogs(pipelineRun.Name, testNamespace)
+		if err != nil {
+			g.GinkgoWriter.Printf("got error fetching PR logs: %v\n", err.Error())
+			continue
+		}
+
+		if err := os.WriteFile(filepath, []byte(pipelineLogs), 0644); err != nil {
+			g.GinkgoWriter.Printf("cannot write to %s: %+v\n", filepath, err)
 		}
 	}
